requirements.txt
streamlit
google-generativeai
requests
python-dotenv
langchain-google-genai
praw
google-generativeai
python-dotenv